+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+
+## next
+
+* Add a right-click menu for the terminal.  It currently allows copy and
+  paste.  [#136](https://github.com/cdepillabout/termonad/pull/136)  Thanks
+  @jecaro!
+
 ## 2.1.0.0
 
 * Add a menu option to set preferences for a running Termonad session.  The preferences you have set are lost when you end the Termonad session. [#130](https://github.com/cdepillabout/termonad/pull/130)  Thanks @jecaro!